Pass-through signin form if already logged in
We shouldn't render the signin form if a user is already logged in. Its
confusing and results in weird behaviour.

For example, if a user is signed in, receives a "You have a new response
" email and follows the URL within the email, they shouldn't need to
sign in again before being redirected to their new response, as the
session already exists.

Fixes https://github.com/mysociety/alaveteli-professional/issues/288
(and also fixes this for non-pro users).