Make request specs behave like production
We already set `show_exceptions` to `true`, which confusingly means that
exceptions are rescued and rails returns a corresponding HTTP status
code. [1]

This commit also sets `consider_all_requests_local` to `false`, so that
we do indeed get the production behaviour of returning user-friendly
error pages instead of a stacktrace.

If there's a specific need, this can be set on a per-spec basis.

[1] https://github.com/rails/rails/issues/29712