squash! Pass-through signin form if already logged in
Disable failing specs for redirecting to external URLs.

The weirdness of our custom error handling solution means these return a
500 status rather than a 404. The 500 is caused by the exception
handling raising the `RouteNotFound`.

We use this behaviour in `do_post_redirect` [1] so I'm confident it
solves the problem, and can be verified in the browser.

I've expanded on what I think is happening in
https://github.com/mysociety/alaveteli/issues/4353.

[1]
https://github.com/mysociety/alaveteli/blob/0.30.0.1/app/controllers/application_controller.rb#L334